Limit themes to built-in presets
