+++ README.md
@@ -52,25 +52,16 @@ Glues offers various storage options to suit your needs:
   - These formats store notes as simple log files, ideal for quick data exports or reading logs.
   - CSV saves data in comma-separated format, while JSON uses JSONL (JSON Lines) format.
 
-### Custom Themes
+### Theme Presets
 
-You can load a custom color theme from a TOML file using the `--theme` option:
+Glues includes several built-in color schemes. Choose one with the `--theme` option:
 
 ```bash
-glues --theme /path/to/theme.toml
+glues --theme pastel
 ```
 
-The file must define all semantic color names with RGB hex values:
-
-```toml
-background = "#000000"
-text = "#ffffff"
-accent = "#ff00ff"
-# ... and so on for each field
-```
-
-You can use the included `sample-theme.toml` file as a starting point and adjust
-the colors to your liking.
+Available presets are `dark`, `light`, and `pastel`. The pastel palette is defined
+in `pastel-theme.toml` and can be used as a template when contributing new themes.
 
 ## Roadmap
