+++ tui/Cargo.toml
@@ -23,5 +23,3 @@ tokio = { version = "1.41.0", features = ["macros", "rt-multi-thread"] }
 arboard = "3.4.1"
 clap = { version = "4.5.4", features = ["derive"] }
 once_cell = "1.19"
-serde = { version = "1.0", features = ["derive"] }
-toml = "0.8"
